Move state db to geth adapter
