+++ go/geth_adapter/state_db.go
@@ -8,7 +8,7 @@
 // On the date above, in accordance with the Business Source License, use of
 // this software will be governed by the GNU Lesser General Public License v3.
 
-package tosca_adapter
+package geth_adapter
 
 import (
 	"github.com/0xsoniclabs/tosca/go/tosca"
